Consider the bed spawn to be in the middle of the block.
Player.get/setBedSpawnLocation() methods round towards -/-
(integer) block coordinates, thereby respawning the player
on the edge of the block. The server will then find a safe
spawn location on the top of a wall, meaning the player
could spawn outside the structure containing the bed.

This effect is particularly bad in conjunction with
EasySigns, where the player could easily spawn outside
of a maze when clicking on a sleep sign.

PlayerRespawnEvent.getRespawnLocation() returns coordinates
that have been offset (0.5, 0.1, 0.5), so that the player
does not have their hitbox against a wall. This version
emulates that approach to the bed spawn location.